rename conftest to utils
