+++ skore/tests/unit/sklearn/plot/roc_curve/test_comparison_cross_validation.py
@@ -11,7 +11,7 @@ from skore import ComparisonReport, CrossValidationReport
 from skore.sklearn._plot.metrics.roc_curve import RocCurveDisplay
 from skore.sklearn._plot.utils import sample_mpl_colormap
 
-from .conftest import check_display_data
+from .utils import check_display_data
 
 
 @pytest.fixture
