+++ skore/tests/unit/sklearn/plot/roc_curve/utils.py
